+++ src/shared/provider/index.ts
@@ -1,1 +1,1 @@
-export { ThemeProvider } from './ThemeProvider';
+export { ThemeProvider } from './theme-provider';
