+++ src/widgets/Header/auth-controls.tsx
